Server implemented PUT request

+++ DB/src/main/resources/migrations/V1_1_0__dummy_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE t_dummy
 (
 
-    id int NOT NULL,
-    num int default 1,
+    id bigint NOT NULL,
+    num bigint default 1,
     CONSTRAINT t_dummy_pk PRIMARY KEY (id)
 );
